+++ tests/fake/structure/migration_directory_2/20150921111111_fourth_add.php
@@ -4,13 +4,13 @@ use Phoenix\Migration\AbstractMigration;
 
 class FourthAdd extends AbstractMigration
 {
-    public function up()
+    public function up(): void
     {
-        
+
     }
-    
-    public function down()
+
+    public function down(): void
     {
-        
+
     }
 }