+++ CHANGELOG.md
@@ -3,10 +3,12 @@
 ### [Unreleased][unreleased]
 
 #### Changed
-- Replaced nette/finder with symfony/finder
+- added return type (:void) for up() and down() methods in AbstractMigration (BC Break - Fix: add return type :void to all migrations)
+- replaced nette/finder with symfony/finder
 
 #### Removed
-- Removed nette/utils
+- removed support for sqlite adapter
+- removed nette/utils
 
 ### [0.13.0] - 2018-03-14
 #### Changed
@@ -15,6 +17,9 @@
 #### Added
 - methods tableExists, tableColumnExists and getTable to migration
 - posibility to add an autoincrement primary key on an existing table
+
+#### Fixed
+- Add a primary key on an existing table
 
 ### [0.12.1] - 2017-12-18
 #### Fixed